% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivw_mvmr.R
\name{ivw_mvmr}
\alias{ivw_mvmr}
\title{ivw_mvmr}
\usage{
ivw_mvmr(r_input, gencov = 0)
}
\arguments{
\item{r_input}{A formatted data frame using the \code{\link[=format_mvmr]{format_mvmr()}} function or an object of class \code{MRMVInput} from \code{\link[MendelianRandomization:mr_mvinput]{MendelianRandomization::mr_mvinput()}}}

\item{gencov}{Retained for interface consistency with the other MVMR functions. The IVW point estimates do not depend on the covariance between the effect of the genetic variants on each exposure, so this argument is not used by \code{ivw_mvmr()}. A value of \code{0} is used by default.}
}
\value{
A dataframe containing MVMR results, including estimated coefficients, their standard errors, t-statistics, and corresponding (two-sided) p-values.
}
\description{
Fits an IVW multivariable Mendelian randomization model using first order weights.
}
\examples{
r_input <- format_mvmr(
  BXGs = rawdat_mvmr[, c("LDL_beta", "HDL_beta")],
  BYG = rawdat_mvmr$SBP_beta,
  seBXGs = rawdat_mvmr[, c("LDL_se", "HDL_se")],
  seBYG = rawdat_mvmr$SBP_se,
  RSID = rawdat_mvmr$SNP
)
ivw_mvmr(r_input)
}
\references{
Sanderson, E., et al., An examination of multivariable Mendelian randomization in the single-sample and two-sample summary data settings. International Journal of Epidemiology, 2019, 48, 3, 713--727. \doi{10.1093/ije/dyy262}
}
\author{
Wes Spiller; Eleanor Sanderson; Jack Bowden.
}
