% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmvinput_to_mvmr_format.R
\name{mrmvinput_to_mvmr_format}
\alias{mrmvinput_to_mvmr_format}
\title{Convert an object of class MRMVInput from the MendelianRandomization package to the MVMR mvmr_format class}
\usage{
mrmvinput_to_mvmr_format(dat)
}
\arguments{
\item{dat}{Object from \code{\link[MendelianRandomization:mr_mvinput]{MendelianRandomization::mr_mvinput()}}.}
}
\value{
Object of class \code{mvmr_format}, the MVMR format
}
\description{
Creates a data.frame with additional class \code{mvmr_format} from an object of class \code{MRMVInput} generated by \code{\link[MendelianRandomization:mr_mvinput]{MendelianRandomization::mr_mvinput()}}.
}
\examples{
if (require("MendelianRandomization", quietly = TRUE)) {
  bx <- as.matrix(rawdat_mvmr[, c("LDL_beta", "HDL_beta")])
  bxse <- as.matrix(rawdat_mvmr[, c("LDL_se", "HDL_se")])
  dat <- MendelianRandomization::mr_mvinput(
    bx = bx,
    bxse = bxse,
    by = rawdat_mvmr$SBP_beta,
    byse = rawdat_mvmr$SBP_se,
    snps = rawdat_mvmr$SNP
  )
  dat <- mrmvinput_to_mvmr_format(dat)
  head(dat)
  class(dat)
}
}
