% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVMR-package.R
\docType{package}
\name{MVMR-package}
\alias{MVMR}
\alias{MVMR-package}
\title{MVMR: MVMR}
\description{
An R package for performing multivariable Mendelian randomization analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WSpiller/MVMR}
  \item \url{https://wspiller.github.io/MVMR/}
  \item \url{https://mrcieu.r-universe.dev/MVMR}
  \item Report bugs at \url{https://github.com/WSpiller/MVMR/issues}
}

}
\author{
\strong{Maintainer}: Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})

Authors:
\itemize{
  \item Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})
  \item Wes Spiller \email{wes.spiller@bristol.ac.uk} (\href{https://orcid.org/0000-0002-8169-5531}{ORCID})
  \item Jack Bowden \email{j.bowden2@exeter.ac.uk} (\href{https://orcid.org/0000-0003-2628-3304}{ORCID})
  \item Eleanor Sanderson \email{eleanor.sanderson@bristol.ac.uk} (\href{https://orcid.org/0000-0001-5188-5775}{ORCID})
}

}
\keyword{internal}
