% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenocov_mvmr.R
\name{phenocov_mvmr}
\alias{phenocov_mvmr}
\title{phenocov_mvmr}
\usage{
phenocov_mvmr(Pcov, seBXGs)
}
\arguments{
\item{Pcov}{A phenotypic matrix using exposures, constructed using individual level exposure data. Columns should be ordered by exposure so as to match \code{format_mvmr}.}

\item{seBXGs}{A matrix containing standard errors corresponding in relation to the gene-exposure association for each SNP.}
}
\value{
A list of covariance matrices with respect to each genetic variant, retaining the ordering in \code{seBXGs}
}
\description{
Uses an external phenotypic covariance matrix and summary data to estimate covariance matrices for estimated effects of individual genetic
variants on each exposure. The phenotypic covariance matrix should be constructed using standardised phenotype measures. The function returns a number of covariance matrices equal to the number of SNPs, where SNP and
row numbers reference ordered exposures.
}
\examples{

\dontrun{
phenocov_mvmr(Pcov, summarydata[,c(3,4)])
}

}
\references{
Sanderson, E., et al., An examination of multivariable Mendelian randomization in the single-sample and two-sample summary data settings. International Journal of Epidemiology, 2019, 48, 3, 713-727. \url{https://dx.doi.org/10.1093/ije/dyy262}
}
\author{
Wes Spiller; Eleanor Sanderson; Jack Bowden.
}
