% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdat_mvmr.R
\docType{data}
\name{rawdat_mvmr}
\alias{rawdat_mvmr}
\title{Raw multivariable MR summary data using lipid fractions as exposures and systolic blood pressure as an outcome.}
\format{
A data frame with 145 rows and 9 variables. A full description of the data is available
by using the command \code{vignette("MVMR")}.
}
\source{
\itemize{
\item \url{http://www.mrbase.org/}
\item \url{https://www.nature.com/articles/ng.2797}
\item \url{https://www.nature.com/articles/ng.3768}
}
}
\usage{
rawdat_mvmr
}
\description{
A dataset containing summary data on 145 genetic variants associated with either
low-density lipoprotein (LDL), high-density lipoprotein (HDL), or triglycerides. Data includes variant rsid numbers,
associations with each lipid fraction, the associations between genetic variants and systolic blood pressure,
and corresponding standard errors.
}
\details{
rawdat_mvmr
}
\author{
Wes Spiller; Eleanor Sanderson; Jack Bowden.
}
\keyword{datasets}
