% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpcov_mvmr.R
\name{snpcov_mvmr}
\alias{snpcov_mvmr}
\title{snpcov_mvmr}
\usage{
snpcov_mvmr(Gs, Xs)
}
\arguments{
\item{Gs}{A matrix or dataframe containing genetic instrument measures. Columns should indicate genetic variant number, with rows representing an observed measure of the genetic variant.}

\item{Xs}{A matrix or dataframe containing exposure measures. Columns should indicate exposure number, with rows representing an observed measure for the given exposure.}
}
\value{
A list of covariance matrices with respect to each genetic variant, retaining the ordering in \code{Gs}
}
\description{
Uses individual level genetic and exposure data to generate covariance matrices for estimated effects of individual genetic
variants on each exposure. The function returns a number of covariance matrices equal to the number of SNPs, where SNP and
row numbers reference ordered exposures.
}
\examples{
\dontrun{
snpcov_mvmr(data[,1:10], data[,11:13])
}
}
\references{
Sanderson, E., et al., An examination of multivariable Mendelian randomization in the single-sample and two-sample summary data settings. International Journal of Epidemiology, 2019, 48, 3, 713--727. \doi{10.1093/ije/dyy262}
}
\author{
Wes Spiller; Eleanor Sanderson; Jack Bowden.
}
