% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivw_mvmr.R
\name{ivw_mvmr}
\alias{ivw_mvmr}
\title{ivw_mvmr}
\usage{
ivw_mvmr(r_input, gencov)
}
\arguments{
\item{r_input}{A formatted data frame using the \code{format_mvmr} function.}

\item{gencov}{Calculating heterogeneity statistics requires the covariance between the effect of the genetic variants on each exposure to be known. This can either be estimated from individual level data, be assumed to be zero, or fixed at zero using non-overlapping samples of each exposure GWAS. A value of \code{0} is used by default.}
}
\value{
An dataframe containing MVMR results, including estimated coefficients, their standard errors, t-statistics, and corresponding (two-sided) p-values.
}
\description{
Fits an IVW multivariable Mendelian randomization model using first order weights.
}
\examples{

r_input <- format_mvmr(
    BXGs = rawdat_mvmr[,c("LDL_beta","HDL_beta")],
    BYG = rawdat_mvmr$SBP_beta,
    seBXGs = rawdat_mvmr[,c("LDL_se","HDL_se")],
    seBYG = rawdat_mvmr$SBP_se,
    RSID = rawdat_mvmr$SNP)
ivw_mvmr(r_input)

}
\references{
Sanderson, E., et al., An examination of multivariable Mendelian randomization in the single-sample and two-sample summary data settings. International Journal of Epidemiology, 2019, 48, 3, 713-727. \url{https://dx.doi.org/10.1093/ije/dyy262}
}
\author{
Wes Spiller; Eleanor Sanderson; Jack Bowden.
}
