% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhet_mvmr.R
\name{qhet_mvmr}
\alias{qhet_mvmr}
\title{qhet_mvmr}
\usage{
qhet_mvmr(r_input, pcor)
}
\arguments{
\item{r_input}{A formatted data frame using the \code{format_mvmr} function.}

\item{pcor}{A phenotypic correlation matrix including the correlation between each exposure included in the MVMR analysis.}
}
\value{
An dataframe containing effect estimates with respect to each exposure. Currently only point estimates are provided.
}
\description{
Fits a multivariable Mendelian randomization model adjusting for weak instruments. The functions requires a
formatted dataframe using the \code{format_mvmr} function, as well a phenotypic correlation matrix \code{pcor}. This should be obtained from individual level
phenotypic data, or constructed as a correlation matrix where correlations have previously been reported.
}
\examples{

qhet_mvmr(r_input,pcor)

}
\references{
Sanderson, E., et al., An examination of multivariable Mendelian randomization in the single-sample and two-sample summary data settings. International Journal of Epidemiology, 2018. [Internet]. 2018;dyy262. Available from: https://dx.doi.org/10.1093/ije/dyy262
}
\author{
Wes Spiller; Eleanor Sanderson; Jack Bowden.
}
